Move DataStoreEvent into persistence.remote
