+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/DataStoreEvent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.service;
+package com.google.android.gnd.persistence.remote;
 
 import java8.util.Optional;
