+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/RemoteDataStore.java
@@ -16,7 +16,6 @@
 
 package com.google.android.gnd.persistence.remote;
 
-import com.google.android.gnd.service.DataStoreEvent;
 import com.google.android.gnd.system.AuthenticationManager.User;
 import com.google.android.gnd.vo.Feature;
 import com.google.android.gnd.vo.FeatureUpdate.RecordUpdate.ResponseUpdate;
